refinement on generator 1D v4 and v10 added: - self attention layer - lr scheduler bug fixes

+++ src/models/generator/modules/GaussianNoise.py
@@ -13,6 +13,6 @@ class GaussianNoise(nn.Module):
             shape = x.shape
             # Generate a NumPy array with the same shape, filled with random values in [0, 1) from gaus distribution
             noise = np.random.rand(*shape)
-            noise = torch.from_numpy(noise).to(device=x.device).float() * self.stddev
+            noise = torch.from_numpy(noise).to(torch.float32).to(device=x.device).float() * self.stddev
             return x + noise
         return x
